#243: Switched to appSetup
